Moved feed 'initisalizing' code to _layout instead of index. This means tht as soon as a user opens any page, there will be a prefetching of data for the feed
